[UPDATE] Login, Logout y bitacaro de bases de datos

+++ src/main/java/utez/edu/mx/practica1/model/person/DaoPerson.java
@@ -1,5 +1,9 @@
 package utez.edu.mx.practica1.model.person;
 
+
+import org.json.JSONArray;
+import org.json.JSONObject;
+import utez.edu.mx.practica1.model.transaction.BeanTransaction;
 import utez.edu.mx.practica1.service.ConnectionDB;
 
 import java.sql.*;
@@ -15,6 +19,7 @@ public class DaoPerson {
 
     public int login(String correo,String contrasena){
        int flag = 0;
+       JSONObject selPer = new JSONObject();
         try{
             con = ConnectionDB.getConnection();
             pstm = con.prepareStatement("SELECT * FROM persona WHERE correo = ? AND contrasena = ?;");
@@ -23,6 +28,14 @@ public class DaoPerson {
             rs = pstm.executeQuery();
 
             if(rs.next()){
+                Long idPersona = rs.getLong("id");
+                selPer.put("accion","login");
+                //QUERY PARA GENERAR LA OPERACIÓN
+                pstm = con.prepareStatement("INSERT INTO json_persona(datos,id_persona,tipoaccion) values(JSON_INSERT('{}','$.json',?),?,?);");
+                pstm.setString(1,selPer.toString());
+                pstm.setLong(2,idPersona);
+                pstm.setString(3,"login");
+                pstm.executeUpdate();
                flag =1 ;
 
             }
@@ -32,6 +45,87 @@ public class DaoPerson {
             ConnectionDB.closeConnections(con,pstm,rs);
         }
         return flag;
+    }
+
+    public boolean signUp(BeanPerson person){
+        JSONObject newPerson = new JSONObject();
+        Boolean flag =false;
+        try{
+            con = ConnectionDB.getConnection();
+            pstm = con.prepareStatement("INSERT INTO persona(nombre, a_paterno, a_materno, edad, sexo, telefono, direccion, " +
+                    "fecha_nacimiento, estado_civil, correo, contrasena,estado) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?,?)", PreparedStatement.RETURN_GENERATED_KEYS);
+            pstm.setString(1,person.getNombre());
+            pstm.setString(2,person.getaPaterno());
+            pstm.setString(3,person.getaMaterno());
+            pstm.setLong(4,person.getEdad());
+            pstm.setString(5,person.getSexo());
+            pstm.setString(6,person.getTelefono());
+            pstm.setString(7,person.getDireccion());
+            pstm.setString(8,person.getFechaNacimiento());
+            pstm.setString(9,person.getEstadoCivil());
+            pstm.setString(10,person.getCorreo());
+            pstm.setString(11,person.getContrasena());
+            pstm.setBoolean(12,person.isEstado());
+
+            //Creación de objeto json para bitacora
+            newPerson.put("nombre",person.getNombre());
+            newPerson.put("aPaterno",person.getaPaterno());
+            newPerson.put("aMaterno",person.getaMaterno());
+            newPerson.put("edad",person.getEdad());
+            newPerson.put("sexo",person.getSexo());
+            newPerson.put("telefono",person.getTelefono());
+            newPerson.put("direccion",person.getDireccion());
+            newPerson.put("fechaNacimiento",person.getFechaNacimiento());
+            newPerson.put("estadoCivil",person.getEstadoCivil());
+            newPerson.put("correo",person.getCorreo());
+            newPerson.put("contrasena",person.getContrasena());
+            newPerson.put("estado",person.isEstado());
+            newPerson.put("accion","signup");
+
+            if(pstm.executeUpdate() == 1){
+                ResultSet lastInsert = pstm.getGeneratedKeys();
+                lastInsert.next();
+                Long lastId = lastInsert.getLong(1);
+                pstm = con.prepareStatement("SELECT * FROM persona WHERE id = ?");
+                pstm.setLong(1, lastId);
+                rs = pstm.executeQuery();
+                lastInsert.close();
+
+                if(rs.next()){
+                    //QUERY PARA GENERAR LA OPERACIÓN
+                    pstm = con.prepareStatement("INSERT INTO json_persona(datos,id_persona,tipoaccion) values(JSON_INSERT('{}','$.json',?),?,?);");
+                    pstm.setString(1, newPerson.toString());
+                    pstm.setLong(2,lastId);
+                    pstm.setString(3,"signup");
+                    pstm.executeUpdate();
+                }
+                flag = true;
+            }
+        }catch (SQLException e){
+            System.out.println("DP_ERR_000R" + e.getMessage());
+        }finally{
+               ConnectionDB.closeConnections(con,pstm);
+        }
+        return flag;
+    }
+
+//Agregar el insert para login y logout para bitacora y hacer vista de registro
+    public void logout(Long id){
+        JSONObject logout = new JSONObject();
+        try{
+            logout.put("accion","logout");
+            con = ConnectionDB.getConnection();
+            pstm = con.prepareStatement("INSERT INTO json_persona(datos,id_persona,tipoaccion) values(JSON_INSERT('{}','$.json',?),?,?);");
+            pstm.setString(1,logout.toString());
+            pstm.setLong(2,id);
+            pstm.setString(3,"logout");
+            pstm.executeUpdate();
+        }catch (SQLException e){
+            System.out.println("DP_ERR_LOGOUT");
+        }finally {
+            ConnectionDB.closeConnections(con,pstm);
+        }
+
     }
 
     public BeanPerson getPersonByLogin(String correo,String contrasena){
@@ -67,9 +161,10 @@ public class DaoPerson {
         return personLog;
     }
     // EncontrarTodos
-    public List<BeanPerson> findAll(){
+    public List<BeanPerson> findAll(Long idPersona){
         List<BeanPerson> listPersons = new ArrayList<>();
         BeanPerson person = null;
+        JSONObject selPer = new JSONObject();
         try{
             con = ConnectionDB.getConnection();
             pstm = con.prepareStatement("SELECT * FROM persona");
@@ -93,8 +188,17 @@ public class DaoPerson {
 
                 listPersons.add(person);
             }
+
+
+            selPer.put("accion","findall");
+            //QUERY PARA GENERAR LA OPERACIÓN
+            pstm = con.prepareStatement("INSERT INTO json_persona(datos,id_persona,tipoaccion) values(JSON_INSERT('{}','$.json',?),?,?);");
+            pstm.setString(1,selPer.toString());
+            pstm.setLong(2,idPersona);
+            pstm.setString(3,"findall");
+            pstm.executeUpdate();
         }catch (Exception e) {
-            System.out.println("DP_ERR_01");
+            System.out.println("DP_ERR_01" + e.getMessage());
         }finally {
             ConnectionDB.closeConnections(con,pstm,rs);
         }
@@ -106,6 +210,7 @@ public class DaoPerson {
     // EncontrarPorId
     public BeanPerson findById(Long id){
         BeanPerson person = null;
+        JSONObject selPer = new JSONObject();
         try {
             con = ConnectionDB.getConnection();
             pstm = con.prepareStatement("SELECT * FROM persona WHERE id = ?");
@@ -127,10 +232,25 @@ public class DaoPerson {
                 person.setCorreo(rs.getString("correo"));
                 person.setEstado(rs.getBoolean("estado"));
                 person.setContrasena(rs.getString("contrasena"));
+
+                Long idPer = rs.getLong("id");
+                selPer.put("id",rs.getLong("id"));
+                selPer.put("nombre",rs.getString("nombre"));
+                selPer.put("aPaterno",rs.getString("a_Paterno"));
+                selPer.put("aMaterno",rs.getString("a_Materno"));
+                selPer.put("correo",rs.getString("correo"));
+                selPer.put("accion","findbyid");
+                //QUERY PARA GENERAR LA OPERACIÓN
+                pstm = con.prepareStatement("INSERT INTO json_persona(datos,id_persona,tipoaccion) values(JSON_INSERT('{}','$.json',?),?,?);");
+                pstm.setString(1,selPer.toString());
+                pstm.setLong(2,idPer);
+                pstm.setString(3,"findbyid");
+                pstm.executeUpdate();
+
             }
 
         } catch (Exception e) {
-            System.out.println("DP_ERR_002");
+            System.out.println("DP_ERR_002" + e.getMessage());
         }finally {
             ConnectionDB.closeConnections(con,pstm);
         }
@@ -140,6 +260,7 @@ public class DaoPerson {
     // Create
     public boolean create(BeanPerson person){
         boolean flag = false;
+        JSONObject newPerson = new JSONObject();
         try{
             con = ConnectionDB.getConnection();
             pstm = con.prepareStatement("INSERT INTO persona(nombre, a_paterno, a_materno, edad, sexo, telefono, direccion, " +
@@ -157,20 +278,42 @@ public class DaoPerson {
             pstm.setString(11,person.getContrasena());
             pstm.setBoolean(12,person.isEstado());
 
-            rs = pstm.getGeneratedKeys();
-            if(pstm.executeUpdate() == 1 && rs.first()){
-                pstm = con.prepareStatement("SELECT * FROM person WHERE id = ?");
-                pstm.setLong(1, rs.getLong(1));
+            //Creación de objeto json para bitacora
+            newPerson.put("nombre",person.getNombre());
+            newPerson.put("aPaterno",person.getaPaterno());
+            newPerson.put("aMaterno",person.getaMaterno());
+            newPerson.put("edad",person.getEdad());
+            newPerson.put("sexo",person.getSexo());
+            newPerson.put("telefono",person.getTelefono());
+            newPerson.put("direccion",person.getDireccion());
+            newPerson.put("fechaNacimiento",person.getFechaNacimiento());
+            newPerson.put("estadoCivil",person.getEstadoCivil());
+            newPerson.put("correo",person.getCorreo());
+            newPerson.put("contrasena",person.getContrasena());
+            newPerson.put("estado",person.isEstado());
 
+            if(pstm.executeUpdate() == 1){
+                ResultSet lastInsert = pstm.getGeneratedKeys();
+                lastInsert.next();
+                Long lastId = lastInsert.getLong(1);
+                pstm = con.prepareStatement("SELECT * FROM persona WHERE id = ?");
+                pstm.setLong(1, lastId);
+                rs = pstm.executeQuery();
+                lastInsert.close();
+                flag = true;
                 if(rs.next()){
                     //QUERY PARA GENERAR LA OPERACIÓN
-                    flag = true;
+                     pstm = con.prepareStatement("INSERT INTO json_persona(datos,id_persona,tipoaccion) values(JSON_INSERT('{}','$.json',?),?,?);");
+                     pstm.setString(1, newPerson.toString());
+                     pstm.setLong(2,lastId);
+                     pstm.setString(3,"create");
+                     pstm.executeUpdate();
                 }
             }
         }catch (SQLException e){
-            System.out.println("DP_ERR_003");
+            System.out.println("DP_ERR_003" + e.getMessage());
         }finally {
-            ConnectionDB.closeConnections(con,pstm);
+            ConnectionDB.closeConnections(con,pstm,rs);
         }
         return flag;
     }
@@ -178,6 +321,7 @@ public class DaoPerson {
     // Update
     public boolean update(BeanPerson person){
         boolean flag = false;
+        JSONObject newPerson = new JSONObject();
         try{
             con = ConnectionDB.getConnection();
             pstm = con.prepareStatement("UPDATE persona SET nombre = ?, a_paterno = ?, a_materno = ?, edad = ?, sexo = ?, telefono = ?, " +
@@ -195,13 +339,32 @@ public class DaoPerson {
             pstm.setBoolean(11,person.isEstado());
             pstm.setLong(12, person.getId());
 
-            if(pstm.executeUpdate() == 1){
-                // QUERY PARA REGISTRAR LA OPERACIÓN
+            //Creación de objeto json para bitacora
+            newPerson.put("id",person.getId());
+            newPerson.put("nombre",person.getNombre());
+            newPerson.put("aPaterno",person.getaPaterno());
+            newPerson.put("aMaterno",person.getaMaterno());
+            newPerson.put("edad",person.getEdad());
+            newPerson.put("sexo",person.getSexo());
+            newPerson.put("telefono",person.getTelefono());
+            newPerson.put("direccion",person.getDireccion());
+            newPerson.put("fechaNacimiento",person.getFechaNacimiento());
+            newPerson.put("estadoCivil",person.getEstadoCivil());
+            newPerson.put("correo",person.getCorreo());
+            newPerson.put("contrasena",person.getContrasena());
+            newPerson.put("estado",person.isEstado());
 
+            if(pstm.executeUpdate() == 1){
+                    //QUERY PARA GENERAR LA OPERACIÓN
+                    pstm = con.prepareStatement("UPDATE json_persona SET datos = JSON_SET('{}','$.json',?),tipoaccion = ? WHERE id_persona = ?;");
+                    pstm.setString(1,newPerson.toString());
+                    pstm.setString(2,"update");
+                pstm.setLong(3,person.getId());
+                    pstm.executeUpdate();
                 flag = true;
             }
         }catch (SQLException e){
-            System.out.println("DP_ERR_004");
+            System.out.println("DP_ERR_004" + e.getMessage());
         }finally {
             ConnectionDB.closeConnections(con,pstm);
         }
@@ -227,16 +390,84 @@ public class DaoPerson {
         boolean flag=false;
         try{
             con = ConnectionDB.getConnection();
-            pstm = con.prepareCall("DELETE FROM persona WHERE id=?");
-
+            pstm = con.prepareCall("DELETE FROM json_persona WHERE id_persona = ?;");
             pstm.setLong(1,id);
+            if(pstm.executeUpdate() == 1){
+                    //QUERY PARA GENERAR LA OPERACIÓN
+                pstm = con.prepareCall("DELETE FROM persona WHERE id=?");
+                pstm.setLong(1,id);
+                   if( pstm.executeUpdate() == 1){
+                       flag = true;
+                   }
 
-            flag = pstm.executeUpdate() == 1;
+            }
+
+
         }catch(SQLException e){
-            System.out.println("DP_ERR_005");
+            System.out.println("DP_ERR_005" + e.getMessage());
         }finally {
             ConnectionDB.closeConnections(con,pstm);
         }
         return flag;
     }
+
+    public List<BeanTransaction> getTransactions () {
+        List<BeanTransaction> transactions = new ArrayList<>();
+        BeanTransaction opera = null;
+        try{
+            con = ConnectionDB.getConnection();
+            pstm = con.prepareStatement("SELECT op.id,CONCAT(P.nombre, ' ', p.a_paterno, ' ', p.a_materno ) as nombre,op.fecha as fechamovimiento,t.nombre as tipo,op.dato_viejo,op.dato_nuevo FROM operacion op join json_persona jp on op.id_jsonpersona = jp.id\n" +
+                    "join persona p on jp.id_persona = p.id\n" +
+                    "join tipo_operacion t on op.id_tipo_operacion = t.id;");
+            rs = pstm.executeQuery();
+
+            while(rs.next()){
+                    opera = new BeanTransaction();
+                opera.setId(rs.getInt("id"));
+                opera.setNombrePersona(rs.getString("nombre"));
+                opera.setFechaMovimiento(rs.getString("fechaMovimiento"));
+                opera.setTipo(rs.getString("tipo"));
+                opera.setDatoViejo(rs.getString("dato_viejo"));
+                opera.setDatoNuevo(rs.getString("dato_nuevo"));
+                transactions.add(opera);
+            }
+
+        }catch (SQLException e){
+            System.out.println("DP_ERR_00TS");
+        }finally {
+            ConnectionDB.closeConnections(con,pstm,rs);
+        }
+        return transactions;
+    }
+
+    public BeanTransaction transactionById(Long id){
+        BeanTransaction opera= null;
+        try {
+            con = ConnectionDB.getConnection();
+            pstm = con.prepareStatement("SELECT op.id,CONCAT(P.nombre, ' ', p.a_paterno, ' ', p.a_materno ) as nombre,op.fecha as fechamovimiento,t.nombre as tipo,op.dato_viejo,op.dato_nuevo FROM operacion op join json_persona jp on op.id_jsonpersona = jp.id\n" +
+                    "join persona p on jp.id_persona = p.id\n" +
+                    "join tipo_operacion t on op.id_tipo_operacion = t.id WHERE op.id = ?;");
+            pstm.setLong(1,id);
+            rs = pstm.executeQuery();
+            if(rs.next()){
+                opera = new BeanTransaction();
+
+                opera.setId(rs.getInt("id"));
+                opera.setNombrePersona(rs.getString("nombre"));
+                opera.setFechaMovimiento(rs.getString("fechaMovimiento"));
+                opera.setTipo(rs.getString("tipo"));
+                opera.setDatoViejo(rs.getString("dato_viejo"));
+                opera.setDatoNuevo(rs.getString("dato_nuevo"));
+                System.out.println("Transaccion " + opera);
+            }
+
+        } catch (Exception e) {
+            System.out.println("DP_ERR_00FT" + e.getMessage());
+        }finally {
+            ConnectionDB.closeConnections(con,pstm);
+        }
+        return opera;
+    }
+
 }
+
